+++ src/biomes/BiomeType.ts
@@ -1,0 +1,5 @@
+
+export enum BiomeType {
+  BuiltIn,
+  Custom
+}
